Added working sortPosition method. Can be called on any column and can optionally be constrained to another column (for example parent_id).
This plays out in both the edit and process stages.

Updated schema.sql for ENUM values of cms_cols to add 'position' as an edit_module option. I am considering ditching the ENUM in general and moving to a some kind of code based config? or plain text entry at worst case.